refactor: change token file name to login
